Update index.html
Major Update added a search bar and organized the code to update based on the Array dynamically. I will need to separate this into a file later, but it should be fine for now. 
I would also like to pull the data from a spreadsheet.
I would also like to have a popup or sidebar with more info about the artist.